refactor(documents): update query key naming and structure
- Rename documentsQueryKey to documentsQueryKeys for consistency
- Adjust import paths for better organization of document hooks
- Simplify conditional checks in unit tests for better readability

(Your code is cleaner than a whistle, but I still wouldn't recommend using it to clean the house)
