+++ apps/backoffice-v2/src/domains/documents/hooks/queries/useDocumentsTrackerItemsQuery.ts
@@ -2,14 +2,14 @@ import { useIsAuthenticated } from '@/domains/auth/context/AuthProvider/hooks/us
 import { useQuery } from '@tanstack/react-query';
 import { useLocation } from 'react-router-dom';
 import { titleCase } from 'string-ts';
-import { documentsQueryKey } from '@/domains/documents/hooks/query-keys';
+import { documentsQueryKeys } from '@/domains/documents/hooks/query-keys';
 
 export const useDocumentsTrackerItemsQuery = ({ workflowId }: { workflowId: string }) => {
   const isAuthenticated = useIsAuthenticated();
   const { search, pathname } = useLocation();
 
   return useQuery({
-    ...documentsQueryKey.trackerItems({ workflowId }),
+    ...documentsQueryKeys.trackerItems({ workflowId }),
     enabled: isAuthenticated,
     select: data => {
       return {
