+++ apps/backoffice-v2/src/domains/documents/hooks/query-keys.ts
@@ -1,7 +1,7 @@
 import { createQueryKeys } from '@lukemorales/query-key-factory';
-import { fetchDocumentsTrackerItems } from '@/domains/documents/hooks/fetchers';
+import { fetchDocumentsTrackerItems } from '@/domains/documents/fetchers';
 
-export const documentsQueryKey = createQueryKeys('documents', {
+export const documentsQueryKeys = createQueryKeys('documents', {
   trackerItems: ({ workflowId }: { workflowId: string }) => ({
     queryKey: ['documents-tracker-items', workflowId],
     queryFn: () => fetchDocumentsTrackerItems({ workflowId }),
